TST added some tests and did a small refactor
