+++ ngmix/priors/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+from ._priors import *
+from ._random import make_rng, srandu
